+++ src/aioquic/quic/connection.py
@@ -1447,6 +1447,7 @@ class QuicConnection:
             cafile=self._configuration.cafile,
             capath=self._configuration.capath,
             cipher_suites=self.configuration.cipher_suites,
+            extra_ca_certs=self._configuration.extra_ca_certs,
             is_client=self._is_client,
             logger=self._logger,
             max_early_data=None if self._is_client else MAX_EARLY_DATA,